refactor: Product Detail data세팅하는 부분 service code 분리(#231)
